+++ common/src/vault/mod.rs
@@ -5,24 +5,19 @@ use crate::{
     supply::SupplyPosition,
     vault::wad::Wad,
 };
+pub use external::*;
 use near_sdk::{
     env,
     json_types::{U128, U64},
     near, require, AccountId, AccountIdRef, Gas, Promise, PromiseOrValue,
 };
-use schemars::{schema::Schema, JsonSchema, SchemaGenerator};
+pub use params::*;
 
+pub mod external;
+pub mod params;
 pub mod wad;
 
 pub type TimestampNs = u64;
-
-pub const DAY_NS: u64 = 86_400_000_000_000;
-pub const YEAR_NS: u64 = 365 * DAY_NS;
-
-pub const MIN_TIMELOCK_NS: u64 = 0;
-pub const MAX_TIMELOCK_NS: u64 = 30 * DAY_NS;
-
-pub const MAX_QUEUE_LEN: usize = 64;
 
 pub type ExpectedIdx = u32;
 pub type ActualIdx = u32;
@@ -218,124 +213,11 @@ impl Restrictions {
     }
 }
 
-#[near_sdk::ext_contract(ext_vault)]
-pub trait VaultExt {
-    // Role and admin
-    fn set_curator(account: AccountId);
-    fn set_is_allocator(account: AccountId, allowed: bool);
-    fn submit_guardian(new_g: AccountId);
-    fn accept_guardian();
-    fn revoke_pending_guardian();
-    fn submit_sentinel(new_s: AccountId);
-    fn accept_sentinel();
-    fn revoke_pending_sentinel();
-    fn set_skim_recipient(account: AccountId);
-    fn set_fees(fees: Fees<U128>);
-    fn accept_fees();
-    fn revoke_pending_fees();
-    fn set_restrictions(restrictions: Option<Restrictions>);
-    fn accept_restrictions();
-    fn revoke_pending_restrictions();
-    fn submit_timelock(new_timelock_ns: U64);
-    fn accept_timelock();
-    fn revoke_pending_timelock();
-
-    // Market config and queues
-    fn submit_cap(market: AccountId, new_cap: U128);
-    fn accept_cap(market: AccountId);
-    fn revoke_pending_cap(market: AccountId);
-    fn submit_cap_group_update(update: CapGroupUpdate);
-    fn accept_cap_group_update(update: CapGroupUpdateKey);
-    fn revoke_pending_cap_group_update(update: CapGroupUpdateKey);
-    fn submit_market_removal(market: AccountId);
-    fn revoke_pending_market_removal(market: AccountId);
-    fn set_supply_queue(markets: Vec<AccountId>);
-    fn set_withdraw_queue(queue: Vec<AccountId>);
-
-    // User flows
-    fn withdraw(amount: U128, receiver: AccountId) -> PromiseOrValue<()>;
-    fn redeem(shares: U128, receiver: AccountId) -> PromiseOrValue<()>;
-    fn execute_next_withdrawal_request() -> PromiseOrValue<()>;
-    fn skim(token: AccountId) -> Promise;
-    fn allocate(weights: AllocationWeights, amount: Option<U128>) -> PromiseOrValue<()>;
-
-    // Views
-    fn get_configuration() -> VaultConfiguration;
-    fn get_total_assets() -> U128;
-    fn get_last_total_assets() -> U128;
-    fn get_total_supply() -> U128;
-    fn get_max_deposit() -> U128;
-    fn convert_to_shares(assets: U128) -> U128;
-    fn convert_to_assets(shares: U128) -> U128;
-    fn preview_deposit(assets: U128) -> U128;
-    fn preview_mint(shares: U128) -> U128;
-    fn preview_withdraw(assets: U128) -> U128;
-    fn preview_redeem(shares: U128) -> U128;
-    fn get_cap_groups() -> Vec<(CapGroupId, CapGroupRecord)>;
-    fn get_fee_anchor_timestamp() -> U64;
-    fn get_fees() -> Fees<U128>;
-    fn get_restrictions() -> Option<Restrictions>;
-}
-
 // Add a 20% buffer to a gas estimate
 #[must_use]
 pub const fn buffer(size: u64) -> Gas {
     Gas::from_tgas((size * 6).div_ceil(5))
 }
-
-// Fetching a position
-const GET_SUPPLY_POSITION: u64 = 4;
-pub const GET_SUPPLY_POSITION_GAS: Gas = Gas::from_tgas(GET_SUPPLY_POSITION);
-
-// Create a withdrawal request
-pub const CREATE_WITHDRAW_REQ_GAS: Gas = buffer(5);
-
-// Balance reads against the underlying NEP-141
-pub const FT_BALANCE_OF_GAS: Gas = Gas::from_tgas(5);
-
-// Execute the next withdrawal request on a market
-const EXECUTE_NEXT_SUPPLY_WITHDRAW_REQ: u64 = 20;
-pub const EXECUTE_NEXT_SUPPLY_WITHDRAW_REQ_GAS: Gas =
-    Gas::from_tgas(EXECUTE_NEXT_SUPPLY_WITHDRAW_REQ);
-
-// Extra gas reserved for post-supply verification callbacks, used in
-// paths where we want a conservative safety margin beyond the base
-// estimate.
-pub const SUPPLY_POST_VERIFY_GAS: Gas = Gas::from_tgas(30);
-
-// Callback gas roots for withdraw/supply orchestration.
-
-// Root budget for callbacks after creating a market-side
-// supply-withdrawal request. Encodes: create request, read supply
-// position and settle withdraw accounting.
-pub const WITHDRAW_CREATE_REQUEST_CALLBACK_GAS: Gas =
-    buffer(EXECUTE_NEXT_SUPPLY_WITHDRAW_REQ + AFTER_EXECUTE_NEXT_SUPPLY_WITHDRAW_REQ);
-
-// Budget for the final "settle" phase of a withdraw execution:
-// reconcile principal and idle_balance, and potentially transition to
-// payout or the next market.
-const AFTER_EXECUTE_NEXT_WITHDRAW: u64 = 5 + 5 + AFTER_SEND_TO_USER;
-pub const WITHDRAW_SETTLE_CALLBACK_GAS: Gas = buffer(AFTER_EXECUTE_NEXT_WITHDRAW);
-
-// Budget for executing the next supply-withdrawal request on a market
-// and fetching the updated supply position before the settle step.
-const AFTER_EXECUTE_NEXT_SUPPLY_WITHDRAW_REQ: u64 =
-    GET_SUPPLY_POSITION + AFTER_EXECUTE_NEXT_WITHDRAW;
-pub const WITHDRAW_EXECUTE_FETCH_POSITION_GAS: Gas = buffer(AFTER_EXECUTE_NEXT_SUPPLY_WITHDRAW_REQ);
-
-const AFTER_SUPPLY_2_READ: u64 = 5;
-pub const SUPPLY_POSITION_READ_CALLBACK_GAS: Gas = buffer(AFTER_SUPPLY_2_READ);
-pub const SUPPLY_AFTER_TRANSFER_CHECK_GAS: Gas = buffer(GET_SUPPLY_POSITION + AFTER_SUPPLY_2_READ);
-
-// NOTE: these are taken after running the contract with the gas report and cieled to next whole TGAS.
-pub const SUPPLY_GAS: Gas = buffer(8);
-pub const ALLOCATE_GAS: Gas = buffer(20);
-pub const WITHDRAW_GAS: Gas = buffer(4);
-pub const EXECUTE_WITHDRAW_GAS: Gas = buffer(9);
-pub const SUBMIT_CAP_GAS: Gas = buffer(3);
-
-const AFTER_SEND_TO_USER: u64 = 5;
-pub const AFTER_SEND_TO_USER_GAS: Gas = Gas::from_tgas(AFTER_SEND_TO_USER);
 
 pub fn require_at_least(needed: Gas) {
     let gas = env::prepaid_gas();
@@ -1080,16 +962,6 @@ pub enum Event {
 pub struct FeeAccrualAnchor {
     pub total_assets: U128,
     pub timestamp_ns: U64,
-}
-
-impl JsonSchema for FeeAccrualAnchor {
-    fn schema_name() -> String {
-        "FeeAccrualAnchor".to_string()
-    }
-
-    fn json_schema(gen: &mut SchemaGenerator) -> Schema {
-        gen.subschema_for::<FeeAccrualAnchor>()
-    }
 }
 
 #[derive(Default)]
